Made Double the new main type instead of Long, Long is still working fine. And yes

+++ src/main/kotlin/data/stack/FixedStack.kt
@@ -4,7 +4,7 @@ import data.registers.IntelRegisters
 import data.registers.intelNames
 import environment.errorsCatchable.ErrorType
 import environment.errorsCatchable.nonBlockError
-import helpers.toLong
+import helpers.toDouble
 import registers
 
 /**
@@ -26,7 +26,7 @@ class FixedStack(private val maxSize: Int) {
 	 */
 	fun push(element: Long) {
 		if (isFull()) {
-			registers.write(intelNames[IntelRegisters.ENSF], false.toLong())
+			registers.write(intelNames[IntelRegisters.ENSF], false.toDouble())
 			nonBlockError(ErrorType.STACK_OVERFLOW)
 
 			return
@@ -43,7 +43,7 @@ class FixedStack(private val maxSize: Int) {
 	 */
 	fun pop(): Long {
 		if (isEmpty()) {
-			registers.write(intelNames[IntelRegisters.ENSF], false.toLong())
+			registers.write(intelNames[IntelRegisters.ENSF], false.toDouble())
 			nonBlockError(ErrorType.STACK_UNDERFLOW)
 			return 0
 		}
@@ -70,7 +70,7 @@ class FixedStack(private val maxSize: Int) {
 	 */
 	fun peek(): Long {
 		if (isEmpty()) {
-			registers.write(intelNames[IntelRegisters.ENSF], false.toLong())
+			registers.write(intelNames[IntelRegisters.ENSF], false.toDouble())
 			nonBlockError(ErrorType.STACK_UNDERFLOW)
 			return 0
