+++ src/main/kotlin/internals/instructions/bitwise/shl.kt
@@ -13,6 +13,6 @@ import registers
  * @throws GeneralBitwiseException If an error occurs during the left shift operation.
  */
 fun Bitwise.shl(operand1: RegisterType, operand2: RegisterType) = call("shl") {
-	registers.read(operand1) shl registers.read(operand2).toInt()
+	registers.read(operand1).toLong() shl registers.read(operand2).toInt()
 
 }
