+++ src/main/kotlin/internals/instructions/controlFlow/jnz.kt
@@ -15,7 +15,7 @@ import vm
  */
 fun ControlFlow.jnz(targetAddress: Int, testRegister: RegisterType): Any = try {
 
-	if (registers.read(testRegister) != 0L) {
+	if (registers.read(testRegister).toLong() != 0L) {
 		vm.pc = targetAddress
 	} else {
 		// I'm not sure why I need this else block
