+++ src/main/kotlin/internals/instructions/dataTransfer/dealloc.kt
@@ -8,5 +8,5 @@ import registers
 
 
 fun DataTransfer.dealloc(memAddress: RegisterType) = call("dealloc") {
-	internalMemory.write(MemoryAddress(registers.read(memAddress)), MemoryValue(null))
+	internalMemory.write(MemoryAddress(registers.read(memAddress).toLong()), MemoryValue(null))
 }
