+++ src/main/kotlin/internals/systemCalls/calls/arrayGet.kt
@@ -8,13 +8,13 @@ import internals.systemCalls.SystemCall
 import registers
 
 fun SystemCall.arrayGet(arrayLocationV: RegisterType, arrayIndexV: RegisterType) = call("arrayGet") {
-	val metaData = internalMemory.read(MemoryAddress(registers.read(arrayLocationV))).value!!
-	val index = registers.read(arrayIndexV)
+	val metaData = internalMemory.read(MemoryAddress(registers.read(arrayLocationV).toLong())).value!!
+	val index = registers.read(arrayIndexV).toLong()
 	if (metaData < index) {
 		errors.InvalidMemoryAddressException(index.toString())
 	}
 	registers.write(
-		RegisterType.R2, internalMemory.read(MemoryAddress(registers.read(arrayLocationV) + 2 + index)).value!!
+		RegisterType.R2, internalMemory.read(MemoryAddress(registers.read(arrayLocationV).toLong() + 2 + index)).value!!
 	)
 
 }
