+++ src/main/kotlin/kilb/Arrays.kt
@@ -11,7 +11,7 @@ class Arrays {
 	 * Like F1 = arrayRef
 	 */
 	fun size() {
-		val meta = registers.read(RegisterType.F1)
+		val meta = registers.read(RegisterType.F1).toLong()
 		val count = internalMemory.read(MemoryAddress(meta + 1)).value!!
 		vm.stackOperations.internalStack.push(count)
 	}
